requirements.txt
streamlit>=1.35
psycopg2-binary>=2.9
pandas>=2.0
plotly>=5.18
anthropic>=0.25
google-generativeai>=0.3.0
python-dotenv>=1.0
